[OP] Move elemwise sum op to tensor fcompute, fix broadcast_to backward, fix roi pooling. (#3476)

+++ src/c_api/c_api_ndarray.cc
@@ -97,7 +97,7 @@ int MXImperativeInvoke(AtomicSymbolCreator creator,
     }
     CHECK(infershape.count(op)) << "Op must have FInferShape registered";
     CHECK(infershape[op](attrs, &in_shapes, &out_shapes));
-    CHECK_EQ(out_shapes.size(), infered_num_outputs);
+    CHECK_EQ(out_shapes.size(), static_cast<size_t>(infered_num_outputs));
 
     std::vector<int>& in_types = ret->arg_types;
     std::vector<int>& out_types = ret->out_types;
@@ -112,7 +112,7 @@ int MXImperativeInvoke(AtomicSymbolCreator creator,
     }
     CHECK(infertype.count(op)) << "Op must have FInferShape registered";
     CHECK(infertype[op](attrs, &in_types, &out_types));
-    CHECK_EQ(out_types.size(), infered_num_outputs);
+    CHECK_EQ(out_types.size(), static_cast<size_t>(infered_num_outputs));
 
     // TODO(piiswrong): infer ctx
     Context ctx;
@@ -221,7 +221,7 @@ int MXImperativeInvoke(AtomicSymbolCreator creator,
             engine::CallbackOnComplete on_complete) {
           std::vector<TBlob> input_blobs, aux_blobs, output_blobs;
           auto atop = auxidx.begin();
-          for (int i = 0; i < ndinputs.size(); ++i) {
+          for (size_t i = 0; i < ndinputs.size(); ++i) {
             if (atop != auxidx.end() && i == *atop) {
               aux_blobs.push_back(ndinputs[i].data());
               ++atop;
